+++ .coveralls.yml
@@ -1,1 +1,1 @@
-repo_token: ewoMYU7OzWIXpAC60E3iaL2JMDq85VJlA
+repo_token: Icus3yWtCgvNX2NlDA0bFmQAU7k1pNb2f
