+++ .gitattributes
@@ -3,3 +3,4 @@ docs/* linguist-vendored=true
 demo/* linguist-vendored=true
 *.html linguist-vendored=true
 *.css linguist-vendored=true
+*.scss linguist-vendored=true
