--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "src/js/FormulaParser"]
-path = src/js/FormulaParser
-url = https://github.com/KennethanCeyer/FormulaParser.git
