+++ index.html
@@ -296,7 +296,7 @@ console.log(result.filterData);</code></pre>
             var data = {"operator":"+","operand1":{"operator":"+","operand1":{"value":"1","type":"unit"},"operand2":{"value":"2","type":"unit"}},"operand2":{"value":"3","type":"unit"}};
 
             // .setData({:data}) can load new expression.
-            //$formula.first().formulize('setData', data);
+            //$formula.first().ui('setData', data);
 
             // .draggable is normal jQuery UI plugin. It support to dragging the element by using mouse pick and move.
             $('.formulize-drop .formulize-drop-items .formulize-custom').draggable({
