--- src/formulize.interface.d.ts
@@ -1,13 +1,0 @@
-import { FormulizeOptions } from './formulize/option.interface';
-
-interface FormulizeFunction {
-    (options: FormulizeOptions): JQuery;
-}
-
-interface FormulizePlugin extends FormulizeFunction { }
-
-declare global {
-    interface JQuery {
-        formulize: FormulizePlugin;
-    }
-}
